Meta+LibWeb: Port ValueID.h/cpp generators to invoke_generator()
